alx_travel_app/requirements.txt

amqp==5.3.1
asgiref==3.9.0
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
kombu==5.5.4
mysqlclient==2.2.7
packaging==25.0
prompt-toolkit==3.0.51
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
